% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gina-T7-funs.R
\name{FunTeam7}
\alias{FunTeam7}
\title{Thin shp file and gg-ify it}
\usage{
FunTeam7(myfile, mytoler = 0.1)
}
\arguments{
\item{myfile}{dataframe of long/lats}

\item{mytoler}{numeric, desired tolerance to thin, default is 0.1}
}
\value{
A tibble for ggplot2
}
\description{
Thin shp file and gg-ify it
}
\examples{
# Take in sf data, use FunTeam7, plot output using ggplot2
data(ausex)
datatib <- FunTeam7(myfile = ausex)
datatib \%>\%
  ggplot(aes(x = long, y = lat, group = group)) +
  geom_polygon(fill = "white", colour = "black", lwd = 1)+
  theme_bw()+
  coord_quickmap()

}
