% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_11.R
\name{team_11}
\alias{team_11}
\title{Convert Shape File to Data Frame}
\usage{
team_11(path=getwd(), tol=0.1)
}
\arguments{
\item{path}{file path to a shape file}

\item{tol}{tolerance value used for thinning the polygon}
}
\value{
A data frame of the geographic information of the polygons.
}
\description{
This function extracts data from a shapefile. Reads in the shapefile and extracts data from nested lists, finally recording them into one data frame with columns: long, lat, group, area, sub and additional information containded by the original shape file.
}
\examples{
 nc<- team_11(system.file("shape/nc.shp", package="sf"),0.1)
}
\author{
Qing He
}
