% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{Thin shp file and gg-ify it}
\usage{
team_7(myfile, mytoler = 0.1)
}
\arguments{
\item{myfile}{dataframe of long/lats}

\item{mytoler}{numeric, desired tolerance to thin, default is 0.1}
}
\value{
A tibble for ggplot2
}
\description{
Thin shp file and gg-ify it
}
\examples{
# Take in sf data, use team_7 function, plot output using ggplot2
exdat <- sf::read_sf(system.file("shape/nc.shp", package = "sf"))
datatib <- team_7(myfile = exdat, mytoler = 0.2)
datatib \%>\%
  ggplot(aes(x = long, y = lat, group = group)) +
  geom_polygon(fill = "white", colour = "black", lwd = 1)+
  theme_bw()+
  coord_quickmap()

}
