% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7fun.R
\name{team_7fun}
\alias{team_7fun}
\title{Thin shp file and gg-ify it}
\usage{
team_7fun(mypath, mytoler = 0.1)
}
\arguments{
\item{mypath}{path to file you want to read}

\item{mytoler}{numeric, desired tolerance to thin, default is 0.1}
}
\value{
A tibble for ggplot2
}
\description{
Thin shp file and gg-ify it
}
\examples{
# Take in sf data, use team_7 function, plot output using ggplot2
datatib <- team_7fun(mypath = system.file("shape/nc.shp", package = "sf"), mytoler = 0.2)
}
