% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_one.R
\name{team_one}
\alias{team_one}
\title{Converting sf file into a dataframe/ tibble}
\usage{
team_one(.file, tolerance=0.1)
}
\arguments{
\item{tolerance}{tolerance value used for thinning the polygon}

\item{file}{read "sf" file}
}
\value{
A dataframe or a tibble of the geographic information of the polygons.
}
\description{
This is a helper fucntion to turn one of these matrices into a single data frame/ tibble with variables long, lat, group and order.
}
\details{
Converting sf format into a data frame
}
\examples{
 team_one(system.file("shape/nc.shp", package="sf"),0.1)
}
